+++ CHANGES.rst
@@ -26,6 +26,9 @@ Bug Fixes
 - Fixed a bug in `~baseband.dada.open` where passing a `squeeze` setting is
   ignored when also passing header keywords in 'ws' mode. [#211]
 
+- Raise an exception rather than return incorrect times for Mark 5B files
+  in which the fractional seconds are not set. [#216]
+
 Other Changes and Additions
 ---------------------------